pyproject.toml
[project]
name = "sm-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.6.2",
    "django>=6.0.1",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "dotenv>=0.9.9",
    "drf-spectacular>=0.29.0",
    "pillow>=12.1.0",
    "redis>=7.1.0",
    "supabase>=2.27.1",
]
